Spin off console_commands as service plugin
